moved Web3Connect into separate folder
1. WalletIntegration folder inside Components
